Remove NumbaFuncBase
