+++ dali/python/nvidia/dali/ops.py
@@ -1271,94 +1271,8 @@ example, ``Compose`` automatically arranges copying the data to GPU memory.
 """
     return op_list[0] if len(op_list) == 1 else _CompoundOp(op_list)
 
-class NumbaFunctionBase(metaclass=_DaliOperatorMeta):
-    def __init__(self, impl_name, run_fn, out_types, in_types, outs_ndim, ins_ndim, batch_processing=False, device='cpu', **kwargs):
-        self._schema = _b.GetSchema(impl_name)
-        self._spec = _b.OpSpec(impl_name)
-        self._device = device
-        self._impl_name = impl_name
-
-        kwargs, self._call_args = _separate_kwargs(kwargs)
-
-        for key, value in kwargs.items():
-            self._spec.AddArg(key, value)
-
-        self.run_fn = run_fn
-        self.out_types = out_types 
-        self.in_types = in_types 
-        self.outs_ndim = outs_ndim 
-        self.ins_ndim = ins_ndim
-        self.num_outputs = len(out_types)
-        self.batch_processing = batch_processing
-        self._preserve = True
-
-    @property
-    def spec(self):
-        return self._spec
-
-    @property
-    def schema(self):
-        return self._schema
-
-    @property
-    def device(self):
-        return self._device
-
-    @property
-    def preserve(self):
-        return self._preserve
-
-    def __call__(self, *inputs, **kwargs):
-        inputs = _preprocess_inputs(inputs, self._impl_name, self._device, None)
-        pipeline = _Pipeline.current()
-        if pipeline is None:
-            _Pipeline._raise_pipeline_required("NumbaFunction operator")
-        if (len(inputs) > self._schema.MaxNumInput() or
-                len(inputs) < self._schema.MinNumInput()):
-            raise ValueError(
-                ("Operator {} expects from {} to " +
-                 "{} inputs, but received {}.")
-                .format(type(self).__name__,
-                        self._schema.MinNumInput(),
-                        self._schema.MaxNumInput(),
-                        len(inputs)))
-        for inp in inputs:
-            if not isinstance(inp, _DataNode):
-                raise TypeError(
-                      ("Expected inputs of type `DataNode`. Received input of type '{}'. " +
-                       "Python Operators do not support Multiple Input Sets.")
-                      .format(type(inp).__name__))
-        op_instance = _OperatorInstance(inputs, self, **kwargs)
-        op_instance.spec.AddArg("run_fn", self.run_fn)
-        op_instance.spec.AddArg("out_types", self.out_types)
-        op_instance.spec.AddArg("in_types", self.in_types)
-        op_instance.spec.AddArg("outs_ndim", self.outs_ndim)
-        op_instance.spec.AddArg("ins_ndim", self.ins_ndim)
-        op_instance.spec.AddArg("device", self.device)
-        op_instance.spec.AddArg("batch_processing", self.batch_processing)
-
-        if self.num_outputs == 0:
-            t_name = self._impl_name + "_id_" + str(op_instance.id) + "_sink"
-            t = _DataNode(t_name, self._device, op_instance)
-            pipeline.add_sink(t)
-            return
-        outputs = []
-
-
-        for i in range(self.num_outputs):
-            t_name = op_instance._name
-            if self.num_outputs > 1:
-                t_name += "[{}]".format(i)
-            t = _DataNode(t_name, self._device, op_instance)
-            op_instance.spec.AddOutput(t.name, t.device)
-            op_instance.append_output(t)
-            pipeline.add_sink(t)
-            outputs.append(t)
-        return outputs[0] if len(outputs) == 1 else outputs
-
 _cpu_ops = _cpu_ops.union({"Compose"})
 _gpu_ops = _gpu_ops.union({"Compose"})
-
 
 _load_ops()
 _load_readers_tfrecord()
